Dia 12 recomeçado

+++ src/view/genero/genero_didatico.php
@@ -13,12 +13,14 @@
                     <div class="subtag">
                         <div class="nome_genero">
                             <h3 class="genero"><?php echo $list->getGenero(); ?></h3>
-                            <a href="/generos/didaticos/editar?id=<?php echo $list->getId()?>" class="superLinks" id="edit">
-                                <i class="fa-solid fa-pen fa-xl" style="color: #ffffff;"></i>
-                            </a>
-                            <a href="/generos/didaticos/excluir?id=<?php echo $list->getId()?>" class="superLinks" id="trash">
-                                <i class="fa-solid fa-trash fa-xl" style="color: #fff;"></i>
-                            </a>
+                            <?php if (array_key_exists('tipo_usuario', $_SESSION) && $_SESSION['tipo_usuario'] == 'funcionário') { ?>
+                                <a href="/generos/didaticos/editar?id=<?php echo $list->getId() ?>" class="superLinks" id="edit">
+                                    <i class="fa-solid fa-pen fa-xl" style="color: #ffffff;"></i>
+                                </a>
+                                <a href="/generos/didaticos/excluir?id=<?php echo $list->getId() ?>" class="superLinks" id="trash">
+                                    <i class="fa-solid fa-trash fa-xl" style="color: #fff;"></i>
+                                </a>
+                            <?php } ?>
                         </div>
                     </div>
                 </div>
